hashx/requirements.txt
aniso8601==7.0.0
appdirs==1.4.3
asgiref==3.2.7
astroid==2.4.0
autopep8==1.5.2
cfgv==3.1.0
colorama==0.4.3
distlib==0.3.0
dj-environ==0.0.1
Django==3.0.8
django-appconf==1.0.4
django-debug-toolbar==2.2
django-environ==0.4.5
django-extensions==2.2.9
django-filter==2.2.0
django-graphql-jwt==0.3.1
django-imagekit==4.0.2
django-rest-framework==0.1.0
djangorestframework==3.11.0
filelock==3.0.12
graphene==2.1.8
graphene-django==2.9.1
graphql-core==2.3.1
graphql-relay==2.0.1
identify==1.4.15
importlib-metadata==1.6.0
inflection==0.3.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
nodeenv==1.3.5
pilkit==2.0
Pillow==7.1.1
pre-commit==2.3.0
promise==2.3
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.5.0
pylint-django==2.0.15
pylint-plugin-utils==0.6
python-decouple==3.3
pytz==2019.3
PyYAML==5.3.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
typed-ast==1.4.1
virtualenv==20.0.18
wrapt==1.12.1
zipp==3.1.0
